# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Gastos",
        "description": "Documentación de la API para el manejo de la aplicación de gastos personales",
        "contact": {
            "name": "Save Money - GastosApp",
            "email": "savemoney@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/gastos/lista": {
            "get": {
                "tags": [
                    "Gastos"
                ],
                "summary": "Obtener lista de gastos del usuario autenticado",
                "operationId": "8a8380ae5d860e0685fedae942e7a1f1",
                "responses": {
                    "200": {
                        "description": "Lista de gastos obtenida con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Listado de gastos obtenido exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_users": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_categoria": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "fecha": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-01-19"
                                                    },
                                                    "monto": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 5000
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Hola, esto es una descripción"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/gastos/guardar": {
            "post": {
                "tags": [
                    "Gastos"
                ],
                "summary": "Crear un nuevo gasto",
                "operationId": "2fb734f89a64e0035755cc7e659fedeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_categoria",
                                    "fecha",
                                    "monto",
                                    "descripcion"
                                ],
                                "properties": {
                                    "id_categoria": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-19"
                                    },
                                    "monto": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción del gasto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gasto creado exitosamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/gastos/editar/{id}": {
            "put": {
                "tags": [
                    "Gastos"
                ],
                "summary": "Actualizar un gasto existente",
                "operationId": "99696f7f4724a66eb51c3e014e2757da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del gasto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_categoria",
                                    "fecha",
                                    "monto",
                                    "descripcion"
                                ],
                                "properties": {
                                    "id_categoria": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-19"
                                    },
                                    "monto": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4500
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Gasto actualizado correctamente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gasto actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gasto actualizado con éxito"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_users": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_categoria": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "fecha": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-01-19"
                                                },
                                                "monto": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 4500
                                                },
                                                "descripcion": {
                                                    "type": "string",
                                                    "example": "Gasto actualizado correctamente"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gasto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/gastos/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Gastos"
                ],
                "summary": "Eliminar un gasto",
                "operationId": "85ede5331b2314364673552df5bc170a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del gasto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gasto eliminado exitosamente"
                    },
                    "404": {
                        "description": "Gasto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}